Speed up build
Move gometalinter installation to an earlier position in Dockerfile.
Before this fix, making a change in one of naksu's source files
invalidated the cached gometalinter installation because it came
after the COPY command copying the source files.

This change allows docker to use the previously cached layer when we
make code changes, so gometalinter won't be downloaded and installed
on every build.