Adjust fonts and text colours used in network status
Now we use the current default system font family on Windows 10 and
Ubuntu. On Ubuntu, the font size also matches the one used in system
GUIs. On Windows it's a little bit bigger than the system font.
Making it smaller would have made it less legible in my opinion.
This should be an ok compromise.